refactoring + fixed date navigation
